post-refactor syntax fixes
