+++ app/__init__.py
@@ -7,16 +7,17 @@ from flask_migrate import Migrate
 from flask_login import LoginManager
 from flask_mail import Mail
 from flask_moment import Moment
+from flask_babel import Babel, lazy_gettext as _l
 
 
 db = SQLAlchemy()
 migrate = Migrate()
 login = LoginManager()
 login.login_view = 'auth.login'
-login.login_message = _l('Please log in to access this page.')
+# login.login_message = _l('Please log in to access this page.')
 
 mail = Mail()
-bootstrap = Bootstrap()
+# bootstrap = Bootstrap()
 moment = Moment()
 babel = Babel()
 
@@ -29,7 +30,6 @@ def create_app(config_class=Config):
   migrate.init_app(app, db)
   login.init_app(app)
   mail.init_app(app)
-  bootstrap.init_app(app)
   moment.init_app(app)
   babel.init_app(app)
   
@@ -56,25 +56,22 @@ def create_app(config_class=Config):
         fromaddr='no-reply@' + app.config['MAIL_SERVER'],
         toaddrs=app.config['ADMINS'], subject='QuickGig Failure',
         credentials=auth, secure=secure)
-        mail_handler.setLevel(logging.ERROR)
-        app.logger.addHandler(mail_handler)
+      mail_handler.setLevel(logging.ERROR)
+      app.logger.addHandler(mail_handler)
 
-        if not os.path.exists('logs'):
-          os.mkdir('logs')
-        file_handler = RotatingFileHandler('logs/quickgig.log',
+      if not os.path.exists('logs'):
+        os.mkdir('logs')
+      file_handler = RotatingFileHandler('logs/quickgig.log',
           maxBytes=10240, backupCount=10)
-        file_handler.setFormatter(logging.Formatter(
+      file_handler.setFormatter(logging.Formatter(
           '%(asctime)s %(levelname)s: %(message)s '
           '[in %(pathname)s:%(lineno)d]'))
-        file_handler.setLevel(logging.INFO)
-        app.logger.addHandler(file_handler)
+      file_handler.setLevel(logging.INFO)
+      app.logger.addHandler(file_handler)
 
-        app.logger.setLevel(logging.INFO)
-        app.logger.info('QuickGig startup')
+      app.logger.setLevel(logging.INFO)
+      app.logger.info('QuickGig startup')
 
     return app
 
-
-
-from app import routes, models, errors
 from app import models
