+++ app/auth/forms.py
@@ -2,6 +2,7 @@ from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField, BooleanField, SubmitField, TextAreaField
 from wtforms.validators import DataRequired, ValidationError, Email, EqualTo, Length
 from app.models import User
+from flask_babel import _, lazy_gettext as _l
 
 class LoginForm(FlaskForm):
   username = StringField('Username', validators=[DataRequired()])
