+++ app/templates/_gig.html
@@ -1,5 +1,5 @@
 
-<!-- <table>
+<table>
   <tr valign="top">
     <td><img src="{{ gig.employer.avatar(36) }}"></td>
 	<td>
@@ -8,13 +8,13 @@
 	<td>{{ gig.employer.username }} needs help with:<br>{{ gig.detail }}
 	</td>
   </tr>
-</table> -->
+</table>
 
 
-<table class="table table-hover">
+<!-- <table class="table table-hover">
         <tr>
             <td width="70px">
-                <a href="{{ url_for('main.user', username=gig.author.username) }}">
+                <a href="{{ url_for('main.user', username=gig.employer.username) }}">
                     <img src="{{ gig.employer.avatar(70) }}" />
                 </a>
             </td>
@@ -23,4 +23,4 @@
                     <a href="{{ url_for('main.user', username=gig.employer.username) }}">
                         {{ gig.employer.username }}
                     </a>
-                {% endset %}
+                {% endset %} -->
