+++ app/templates/auth/login.html
@@ -21,6 +21,6 @@
     <p>{{ form.remember_me() }} {{ form.remember_me.label }}</p>
     <p>{{ form.submit() }}</p>
   </form>
-  <p>Forgot Your Password? <a href="{{url_for('password_reset_request') }}">Click to Reset It</a></p>
-  <p>New to QuickGig? <a href="{{ url_for('signup') }}">Click here to sign up!</a></p>
+  <p>Forgot Your Password? <a href="{{url_for('auth.password_reset_request') }}">Click to Reset It</a></p>
+  <p>New to QuickGig? <a href="{{ url_for('auth.signup') }}">Click here to sign up!</a></p>
 {% endblock %}
