+++ config.py
@@ -1,7 +1,7 @@
 import os
 from dotenv import load_dotenv
 basedir = os.path.abspath(os.path.dirname(__file__))
-load dotenv(os.path.join(basedir, '.env'))
+load_dotenv(os.path.join(basedir, '.env'))
 
 class Config(object):
   SECRET_KEY = os.environ.get('SECRET_KEY') or 'you-will-never-guess'
