+++ quickgig.py
@@ -1,8 +1,8 @@
-from app import create_app, db, cli
+from app import create_app, db
 from app.models import User, Gig
 
 app = create_app()
-cli.register(app)
+#cli.register(app)
 
 @app.shell_context_processor
 def make_shell_context():
